requirements.txt
beautifulsoup4==4.9.3
Cython==0.29.14
cmake==3.18.2.post1
decorator==4.4.2
Django==3.2.6
django-cors-headers
h5py
ffmpeg
imageio==2.9.0
imageio-ffmpeg==0.3.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
imutils==0.5.3
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.19.5
oauth==1.0.1
opencv-python==3.4.4.19
opencv-python-headless==4.5.1.48
openpyxl==3.0.6
pika==1.3.0
packaging==20.4
matplotlib==3.3.2
pandas==1.1.5
Pillow==7.2.0
protobuf==3.18.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Shapely==1.7.1
celery==5.1.2
requests==2.26.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
imagezmq==1.1.1
django-timezone-field==4.2.1
flower==1.0.0
djangorestframework==3.12.4
amqp==5.0.6
psycopg2-binary==2.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.0.1




